Even though we don't have range checks, this is the more appropriate type

+++ rtl/tbblue.pas
@@ -5,12 +5,12 @@
 (**
  * Gets the current value of the given Next register.
  *)
-function GetNextReg(Number: Integer): Integer; register; external 'nxt_getreg';
+function GetNextReg(Number: Byte): Byte; register; external 'nxt_getreg';
 
 (**
  * Sets the given Next register to the given value.
  *)
-procedure SetNextReg(Number, Value: Integer); register; external 'nxt_setreg';
+procedure SetNextReg(Number, Value: Byte); register; external 'nxt_setreg';
 
 (**
  * Gets the current CPU speed, with 0..3 meaning 3.5 MHz, 7 MHz, 14 MHz and
